backend/requirements.txt
fastapi>=0.88.0
uvicorn>=0.20.0
pydantic>=1.10.2
textblob>=0.17.1
python-dotenv>=0.21.0
pymongo>=4.3.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
email-validator>=2.0.0
